@@ -1,7 +1,9 @@
 fn main() -> Result<(), Box<dyn std::error::Error>> {
-    let url: String = nekoslife::get(nekoslife::SfwCategory::Waifu)?;
+    let url: String = nekoslife::blocking_get(
+        nekoslife::Category::Waifu,
+    )?;
 
-    println!("{}", url);
+    println!("{url}");
 
     Ok(())
 }
